requirements.txt
Django>=3.0.7
django-import-export>=2.3.0
mysqlclient>=1.4.6
pyexcel-io>=0.5.20
pyexcel-xls>=0.5.8
pyexcel-xlsx>=0.5.8
django-crispy-forms
django-heroku
gunicorn
requests
django-ckeditor
channels
channels
channels-redis
django-extensions
graphviz